// src/app/views/models/missionManagerialAbility.model.ts
import {ManagerialAbilityModel} from './managerialAbility.model';

export class MissionManagerialAbilityModel {
    id: number;
    missionManagerialAbilitiesGrade: number;
    missionManagerialAbilitiesReliability: number;
    managerialAbilities: ManagerialAbilityModel;
    missionId: number;
    managerialAbilityId: number;
}