// src/app/views/models/trainings.model.ts
export class TrainingsModel {
    id: number;
    trainingTitle: string;
    trainingStartDate: string;
    trainingEndDate: string;
    trainingLevelId: number;
    schoolId: number;
    candidateId: number;
    trainingLevelTitle: string;
    schoolName: string;
    missionId: number;
}